recsys/requirements.txt
black==26.1.0
click==8.3.1
mypy_extensions==1.1.0
numpy==2.4.2
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytokens==0.4.1